+++ dflib/src/main/java/com/nhl/dflib/sort/IndexSorter.java
@@ -5,37 +5,36 @@ import com.nhl.dflib.DataFrame;
 import com.nhl.dflib.Series;
 import com.nhl.dflib.row.DataFrameRowProxy;
 import com.nhl.dflib.row.RowProxy;
-import com.nhl.dflib.series.ArraySeries;
-import com.nhl.dflib.series.IndexedSeries;
+import com.nhl.dflib.series.IntIndexedSeries;
+import com.nhl.dflib.series.IntSeries;
 
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.function.Supplier;
 
 public class IndexSorter {
 
     private DataFrame dataFrame;
-    private Supplier<Integer[]> indexBuilder;
+    private Supplier<int[]> indexBuilder;
 
     public IndexSorter(DataFrame dataFrame) {
         this.dataFrame = dataFrame;
         this.indexBuilder = () -> rowNumberSequence(dataFrame.height());
     }
 
-    public IndexSorter(DataFrame dataFrame, Series<Integer> rangeToSort) {
+    public IndexSorter(DataFrame dataFrame, IntSeries rangeToSort) {
         this.dataFrame = dataFrame;
 
         // copy range to avoid modification of the source list
         this.indexBuilder = () -> {
             int len = rangeToSort.size();
-            Integer[] data = new Integer[len];
+            int[] data = new int[len];
             rangeToSort.copyTo(data, 0, 0, len);
             return data;
         };
     }
 
-    protected static Integer[] rowNumberSequence(int h) {
-        Integer[] rn = new Integer[h];
+    protected static int[] rowNumberSequence(int h) {
+        int[] rn = new int[h];
         for (int i = 0; i < h; i++) {
             rn[i] = i;
         }
@@ -43,31 +42,31 @@ public class IndexSorter {
         return rn;
     }
 
-    public Series<Integer> sortIndex(Comparator<RowProxy> comparator) {
+    public IntSeries sortIndex(Comparator<RowProxy> comparator) {
         // make sure 'mutableIndex' is not visible outside this method as we are going to modify it,
         // so obtain it via the supplier right on the spot
-        Integer[] mutableIndex = indexBuilder.get();
+        int[] mutableIndex = indexBuilder.get();
 
-        Comparator<Integer> rowComparator = rowIndexComparator(comparator);
+        IntComparator rowComparator = rowIndexComparator(comparator);
 
         // note - mutating passed index
-        Arrays.sort(mutableIndex, rowComparator);
-        return new ArraySeries<>(mutableIndex);
+        IntTimSort.sort(mutableIndex, rowComparator);
+        return new IntSeries(mutableIndex);
     }
 
     public DataFrame sort(Comparator<RowProxy> comparator) {
-        Series<Integer> sortedIndex = sortIndex(comparator);
+        IntSeries sortedIndex = sortIndex(comparator);
 
         int width = dataFrame.width();
         Series<?>[] newColumnsData = new Series[width];
         for (int i = 0; i < width; i++) {
-            newColumnsData[i] = new IndexedSeries<>(dataFrame.getColumn(i), sortedIndex);
+            newColumnsData[i] = new IntIndexedSeries<>(dataFrame.getColumn(i), sortedIndex);
         }
 
         return new ColumnDataFrame(dataFrame.getColumnsIndex(), newColumnsData);
     }
 
-    private Comparator<Integer> rowIndexComparator(Comparator<RowProxy> rowComparator) {
+    private IntComparator rowIndexComparator(Comparator<RowProxy> rowComparator) {
         DataFrameRowProxy p1 = new DataFrameRowProxy(dataFrame);
         DataFrameRowProxy p2 = new DataFrameRowProxy(dataFrame);
         return (i1, i2) -> rowComparator.compare(p1.rewind(i1), p2.rewind(i2));