+++ dflib/src/main/java/com/nhl/dflib/sort/IntComparator.java
@@ -1,4 +1,4 @@
-package com.nhl.dflib.collection;
+package com.nhl.dflib.sort;
 
 /**
  * @since 0.6